Fix plain:maca input not working
Fix the plain:maca input type not working on Linux (and possibly also on Windows, untested). The issue was caused by the inability of redirecting maca-analyse output to file - redirecting using shell redirections is only possible if the program is ran inside shell, but Process.exec does not provide a shell, which caused the file to not be saved, and thus causing an exception. After replacing with ProcessBuilder, both wcrft and maca work as inteded.